+++ example/ios/Runner/AppDelegate.swift
@@ -1,20 +1,18 @@
-import UIKit
 import Flutter
+import UIKit
 import MiniPlengi
 
 @UIApplicationMain
-@objc class AppDelegate: FlutterAppDelegate, PlaceDelegate {
+@objc class AppDelegate: FlutterAppDelegate {
   override func application(
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
   ) -> Bool {
     GeneratedPluginRegistrant.register(with: self)
-      Plengi.initialize(clientID: "loplat", clientSecret: "loplatsecret")
+      // ********** 중간 생략 ********** //
+    Plengi.initialize(clientID: "loplat", clientSecret: "loplatsecret")
+               
       
     return super.application(application, didFinishLaunchingWithOptions: launchOptions)
   }
-    
-    func responsePlaceEvent(_ plengiResponse: PlengiResponse) {
-        print(plengiResponse)
-    }
 }